AWS Lambda Deployment Package
This package was used for deploying codes to Lambda. It includes Python packages, my code, and two sample images for text detection. Since this is still in the development stage, you need to set up an S3 bucket, Amazon Textract, and AWS System Manager - parameter store on your side In order to use the application.